refactor: separate interface istory in file
