+++ src/types/IStory.ts
@@ -1,0 +1,6 @@
+export interface IStory {
+	id: number;
+	title: string;
+	text: string;
+	imgSrc: string;
+}
